+++ app/src/main/java/com/example/mapsactivity/MapsActivity.kt
@@ -5,6 +5,9 @@ import android.location.Location
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.app.ActivityCompat
+import com.example.mapsactivity.data.Store
+import com.example.mapsactivity.data.Stores
+import com.example.mapsactivity.data.StoresByGeo
 import com.google.android.gms.location.FusedLocationProviderClient
 import com.google.android.gms.location.LocationServices
 import com.google.android.gms.maps.CameraUpdateFactory
@@ -14,6 +17,12 @@ import com.google.android.gms.maps.SupportMapFragment
 import com.google.android.gms.maps.model.LatLng
 import com.google.android.gms.maps.model.Marker
 import com.google.android.gms.maps.model.MarkerOptions
+import com.google.gson.Gson
+import com.google.gson.GsonBuilder
+import com.google.gson.JsonParser
+import com.google.gson.reflect.TypeToken
+import okhttp3.*
+import java.io.IOException
 
 class MapsActivity : AppCompatActivity(), OnMapReadyCallback, GoogleMap.OnMarkerClickListener {
 
@@ -35,6 +44,57 @@ class MapsActivity : AppCompatActivity(), OnMapReadyCallback, GoogleMap.OnMarker
                 arrayOf(android.Manifest.permission.ACCESS_FINE_LOCATION), LOCATION_PERMISSION_REQUEST_CODE)
             return
         }
+    }
+
+    // Json Parser 기능 ( 현재 위도, 경도 값을 받아서 공공데이터 받아오기 )
+    fun fetchJson(location: Location){
+        println("데어터를 가져 오는 중...")
+//        val key = "Your-key"
+
+        // maskApi 링크로 변경함
+        val url = "https://8oi9s0nnth.apigw.ntruss.com/corona19-masks/v1"
+
+        // 임의로 현재 위치로 설정
+        val query = "/storesByGeo/json?"+ "lat=37.500148776257376&lng=127.02741476091066&m=1000";
+        println("--------query---------\n" + url + query)
+        val request = Request.Builder().url(url + query).build()
+        val client = OkHttpClient()
+        client.newCall(request).enqueue(object: Callback {
+            override fun onResponse(call: Call, response: Response) {
+                val body = response?.body()?.string()
+
+                if(body != null) {
+                    println("--------body---------")
+                    println(body)
+                } else {
+                    println("error!")
+                }
+
+                //파싱 - 이렇게 가져온 데이터를 모델오브젝트로 변환해 줘야 한다.
+                val gson = GsonBuilder().create()
+                val parser = JsonParser()
+                //루트 object와 경로를 찾아서 설정해 주는데 이부분에서 개념이 안 잡혀서 헤메다. 원하는 데이터가 속에 속에 들어 있어서...
+                val rootObj = parser.parse(body.toString())
+                    .getAsJsonObject().get("stores")
+                //여기까진 제대로 되고
+                val type = object : TypeToken<List<Store>>() {}.type
+                val storesByGeo = gson.fromJson<List<Store>>(rootObj, type)
+
+//                val storesByGeo =  gson.fromJson<List<Store>>(rootObj, StoresByGeo::class.java)
+                //썸네일을 위한 추가 작업
+                println("--------store[0]---------")
+                println(storesByGeo.get(0).name)
+
+                //백그라운드에서 돌기 때문에 메인UI로 접근할 수 있도록 해줘야 한다.
+                runOnUiThread {
+                    //어답터 설정
+//                    my_recycler_view.adapter = RecyclerViewAdapter(books)
+                }
+            }
+            override fun onFailure(call: Call, e: IOException) {
+                println("리퀘스트 실패")
+            }
+        })
     }
 
     // 새로운 핀 생성 (아직 수정 더 해야함)
@@ -77,6 +137,7 @@ class MapsActivity : AppCompatActivity(), OnMapReadyCallback, GoogleMap.OnMarker
         // 1
         map.isMyLocationEnabled = true
 
+
         // 2
         fusedLocationClient.lastLocation.addOnSuccessListener(this) { location ->
             // 마지막으로 알려준 위치를 갖는다. 가끔 null값이 있을 수도 있음
@@ -86,6 +147,7 @@ class MapsActivity : AppCompatActivity(), OnMapReadyCallback, GoogleMap.OnMarker
                 val currentLatLng = LatLng(location.latitude, location.longitude)
                 map.animateCamera(CameraUpdateFactory.newLatLngZoom(currentLatLng, 16f))
             }
+            fetchJson(lastLocation);
         }
 
         // Add a marker in Seoul and move the camera (처음 서울 기반 zoom)
