--- app/src/main/java/com/example/mapsactivity/data/StoresByGeo.kt
@@ -1,6 +1,0 @@
-package com.example.mapsactivity.data
-
-data class StoresByGeo(
-    val count: Int,
-    val stores: List<Store>
-)
